requirements.txt
PyYAML==3.12
simplejson==3.13.2
pymarketcap==3.3.145
tabulate==0.8.2
#pprint==0.1
#docopt==0.6.2